% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ahat.plot.R
\name{ahat.plot}
\alias{ahat.plot}
\title{Function to make the Ahat plot for genetic enrichment analyses
Builds on grs.summary function from package gtx}
\usage{
ahat.plot(x, genderdiffvector)
}
\arguments{
\item{x}{Input data frame of results from grs.summary at different Pval cutoffs.}

\item{genderdiffvector}{Optional list of pvalue for difference in ahat by sex.}
}
\description{
Function to make the Ahat plot for genetic enrichment analyses
Builds on grs.summary function from package gtx
}
\seealso{
gtx, grs.plot
}

