% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelematch.R
\name{allelematch}
\alias{allelematch}
\title{Function to get logical vector for if alleles match}
\usage{
allelematch(x, y)
}
\arguments{
\item{x, y}{Vectors of alleles}
}
\value{
output Logical vector of length x
}
\description{
Function to return logical vector as response to if alleles in
two vector of alleles match. Takes into account A/T and C/G matching
but does NOT account for snps with only A/T or C/G alleles.
}
\examples{
Match=allelematch(scd$Allele1.scd, scd$Allele1.cad)
}
