% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2num.R
\name{factor2num}
\alias{factor2num}
\title{Function to convert vector to class numeric}
\usage{
factor2num(x)
}
\arguments{
\item{x}{An input vector to be converted to class numeric}
}
\value{
output Numeric vector with non-numeric values replaced with NA
}
\description{
Function to convert vector to class numeric
}
\examples{
test=factor(c("1","4","ABC","5"))
print(as.numeric(test))
print(factor2num(test))
}
\keyword{factor}
\keyword{numeric}
