% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastread.R
\name{fastread}
\alias{fastread}
\title{Function to read in large dataframes faster}
\usage{
fastread(filename)
}
\arguments{
\item{filename}{Name of file to be read in}
}
\description{
Function to read in large dataframes faster
}
\seealso{
gtx, grs.plot
}
