% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/get.make.groups.R
\name{get.make.groups}
\alias{get.make.groups}
\title{Updated version of make.groups}
\usage{
get.make.groups(...)
}
\arguments{
\item{...}{A vector with name of dataframes to be merged}
}
\description{
Updated version of make.groups
}
\seealso{
lattice
}

