% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getR.R
\name{getR}
\alias{getR}
\title{Function to get sum of probes A+B from Affy 6.0 array}
\usage{
getR(x)
}
\arguments{
\item{x}{Matrix with rows as samples and columns as probe intensities}
}
\value{
output Matrix with A+B probe intensities; same number of
rows as input, half as many columns
}
\description{
Function to return A+B probe intensities from matrix with 
A and B intensities
}
\examples{
test2=getR(test)
}

