% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grs.indivbeta.R
\name{grs.indivbeta}
\alias{grs.indivbeta}
\title{Function to get adjusted betas from package gtx}
\usage{
grs.indivbeta(w, b, s, outputname = NULL)
}
\arguments{
\item{w}{Coefficients for the risk score.}

\item{b}{Aligned beta coefficients in the testing dataset, of same 
length as w.}

\item{s}{Standard errors for â, of same length as  w and b.}

\item{outputname}{Optional argument to name of columns of output
dataset. If NULL (default), output matrix has column names ahat
and aSE.}
}
\description{
Function to get betas for individual SNPs used to calculate
ahat from function grs.summary in package gtx
}
\seealso{
gtx, grs.plot
}

