% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.character0.R
\name{is.character0}
\alias{is.character0}
\title{Function to check if object is of class character
and of length 0}
\usage{
is.character0(x)
}
\arguments{
\item{x}{An input vector to be checked}
}
\value{
output Logical output
}
\description{
Function to check if object is of class character
and of length 0
}
\keyword{character}
\keyword{zero}
