% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.integer.R
\name{is.integer0}
\alias{is.integer0}
\title{Function to check if object is an integer of length 0}
\usage{
is.integer0(x)
}
\arguments{
\item{x}{An input vector to be checked}
}
\value{
output Logical output
}
\description{
Function to check if object is an integer of length 0
}
\keyword{integer}
\keyword{zero}

