% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeclumps.all.R
\name{makeclumps.all}
\alias{makeclumps.all}
\title{Function to make the SNP clumps at different Pvalue cutoffs}
\usage{
makeclumps.all(trait, suffix)
}
\arguments{
\item{trait}{Name of trait for which SNPs are being pruned
for eg. QT}

\item{suffix}{Suffix of file names, often name of analysis files
for eg. CAD}
}
\description{
Function to make the SNP clumps at different Pvalue cutoffs
}
\seealso{
gtx, grs.plot
}
