% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeclumps.CAD.R
\name{makeclumps.CAD}
\alias{makeclumps.CAD}
\title{Function to make the SNP clumps at different Pvalue cutoffs}
\usage{
makeclumps.CAD(trait)
}
\arguments{
\item{trait}{Trait to compare CAD results to}
}
\description{
Function to make the SNP clumps at different Pvalue cutoffs
}
\seealso{
gtx, grs.plot
}
