% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAplot.R
\name{PCAplot}
\alias{PCAplot}
\title{Plots grid of PCs}
\usage{
PCAplot(data = data, colorvector = NULL, numPCs = 8, pointsize = 6,
  colorpalette = NULL)
}
\arguments{
\item{data}{dataframe with PCs and colorvector}

\item{colorvector}{name of column by which points should be colored}

\item{numPCs}{how many PCs should be plotted, must be even number (default=8)}

\item{pointsize}{The size of the point for each datapoint}

\item{colorpalette}{Character vector of colors, must be without names}

\item{label_by}{name of column by which points should be labeled}
}
\value{
a ggplot2 plot object
}
\description{
Plots grid of PCs
}
