% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permp.R
\name{permp}
\alias{permp}
\title{Fuction to compute Pval from given null distribution}
\usage{
permp(testnull, testdata)
}
\arguments{
\item{testnull}{Vector of null pval distribution}

\item{testdata}{Vector of pval from dataset of interest}
}
\value{
output Vector of permuted pvalues
}
\description{
Fuction to compute Pval from given null distribution
}

