% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.identified.R
\name{read.identified}
\alias{read.identified}
\title{Plots grid of PCs}
\usage{
read.identified(data)
}
\arguments{
\item{data}{name of file to be read in}
}
\value{
a data frame
}
\description{
Plots grid of PCs
}
