% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wideScreen.R
\name{wideScreen}
\alias{wideScreen}
\title{Function to increase width of R on linux}
\usage{
wideScreen(howWide = as.numeric(strsplit(system("stty size", intern = T),
  " ")[[1]])[2])
}
\value{
output Width of the window is changed
}
\description{
Function to increase width of R on linux
}
