% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeclumps.custom.MAF.R
\name{makeclumps.custom.MAF}
\alias{makeclumps.custom.MAF}
\title{Function to make the SNP clumps at different Pvalue cutoffs}
\usage{
makeclumps.custom.MAF(trait, clumpfile, outfilename, MAF = 0.05)
}
\arguments{
\item{trait}{Trait to compare SCD results to}

\item{clumpfile}{File with rsIDs and pvalues for
trait of interest. Header MUST be SNP and pvalue}

\item{outfilename}{Filename for all output files,
will be followed by alpha cutoff value}

\item{MAF}{Maf cutoff to be used for filtering}
}
\description{
Function to make the SNP clumps at different Pvalue cutoffs
}
\seealso{
gtx, grs.plot
}

