% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeclumps.custom.R
\name{makeclumps.custom}
\alias{makeclumps.custom}
\title{Function to make the SNP clumps at different Pvalue cutoffs}
\usage{
makeclumps.custom(trait, clumpfile, outfilename)
}
\arguments{
\item{trait}{Trait to compare SCD results to}

\item{clumpfile}{File with rsIDs and pvalues for
trait of interest. Header MUST be SNP and pvalue
#param outfilename Filename for all output files,
will be followed by alpha cutoff value}
}
\description{
Function to make the SNP clumps at different Pvalue cutoffs
}
\seealso{
gtx, grs.plot
}

