% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeclumps.R
\name{makeclumps}
\alias{makeclumps}
\title{Function to make the SNP clumps at different Pvalue cutoffs}
\usage{
makeclumps(trait)
}
\arguments{
\item{trait}{Trait to compare SCD results to}
}
\description{
Function to make the SNP clumps at different Pvalue cutoffs
}
\seealso{
gtx, grs.plot
}
